- Containers can be informed when they are required to reload, for example   if new root resource classes have been added to removed, or class file   definitions have changed while a VM is running.   A com.sun.ws.rest.spi.container.ContainerNotifier instance may be   registered by adding a property to the ResourceConfig with the property name   "com.sun.ws.rest.spi.container.ContainerNotifier" and a property value   that is the instance of the ContainerNotifier.   The ContainerNotifier instance may then notifier registered listeners   when reloading is required.
JSPTemplateProcessor now a service provider so will always be registered regardless of the resource configuration.

In service provider constructors access a dependent class. This will ensure the component will not be loaded in cases where no byte code verification is performed or there is some strange kind of class loading going on.

git-svn-id: https://svn.java.net/svn/jersey~svn/trunk@933 2fda5173-fe30-0410-90bd-98ff5453417f