This transaction provides an optional WadlGenerator that examines any JAXB classes that are attached to the Resources in an application in order that a XML Schema can be generated and attached to the WADL.
In order to achieve this another method has been added to the WadlGenerator interface so that a generator can provide external grammar files. In order to make the authoring of grammar generators far less complicated the life cycle of the WadlGenerator has been
modified so that it only lives for the lenght of the generation. Also most of the generation work is now encapsulated in the WadlApplicationContext which should make it much easier to cache the result.

This require so other changes because a WadlGeneratorConfig is now passed around rather than an instance of the WadlGenerator. This class as been modified to just take class
references rather than instances of classes so that the lifecycle can be controlled.

When installed the schemas are laid out as follows:

%root%/application.wadl/xsd0.xsd etc.

Both the root WADL and the WADL generated using OPTIONS for a particular resource link
back to this location.

git-svn-id: https://svn.java.net/svn/jersey~svn/branches/gdavison-wadl-grammar-generation@5164 2fda5173-fe30-0410-90bd-98ff5453417f